Split monolithic source into smaller cpp files
